// src/public_pages/carrier/shared/model/JobCircularModel.ts
export class JobCircularModel {
  JobCircularId: number = 0;
  JobPosition: string = "";
  OrganizationName: string = "";
  TotalNumberOfVacancy: string = "";
  JobContext: string = "";
  JobResponsibility: string = "";
  EmploymentStatus: string = "";
  EducationalRequirements: string = "";
  AdditionalRequirements: string = "";
  Religion: string = "";
  Age: string = "";
  Gender: string = "";
  JobLocation: string = "";
  Salary: string = "";
  CompensationAndOtherBenefits: string = "";
  ApplicationDeadline: string = "";
  PublishedDate: string = "";
  ContactInfo: string = "";
  JobCircularStatus: boolean = false;
  CreatedBy: number = 0;
  UpdatedBy: string = "";
  CreatedAt: string = "";
  UpdatedAt: string = "";
  Remarks: string = "";
  ExperienceRequirements: string = "";
  ReligionName: string = "";
  GenderName: string = "";
  NumberOfJobApplications: number = 0;
}